added GeniaChunkTokenizerDataWriterCPE.xml to list of files not to remove in tearDown
git-svn-id: https://cleartk.googlecode.com/svn/trunk@352 2f4ce1e4-9571-11dd-93b2-6d14b300b761

+++ test/src/org/cleartk/chunk/ChunkHandlerTests.java
@@ -63,7 +63,8 @@ public class ChunkHandlerTests {
 	@After
 	public void tearDown() throws Exception {
 		for (File file: this.outputDir.listFiles()) {
-			if (!file.getName().equals("model.jar")) {
+			if (!file.getName().equals("model.jar") &&
+				!file.getName().equals("GeniaChunkTokenizerDataWriterCPE.xml")) {
 				file.delete();
 			}
 		}
